Update core, modules and theme to latest releases

+++ site/sites/all/themes/corporateclean/template.php
@@ -113,7 +113,7 @@ if (theme_get_setting('slideshow_js','corporateclean')):
 	
 	//Initialize slideshow using theme settings
 	$effect=theme_get_setting('slideshow_effect','corporateclean');
-	$effect_time=theme_get_setting('slideshow_effect_time','corporateclean')*1000;
+	$effect_time= (int) theme_get_setting('slideshow_effect_time','corporateclean')*1000;
 	$slideshow_randomize=theme_get_setting('slideshow_randomize','corporateclean');
 	$slideshow_wrap=theme_get_setting('slideshow_wrap','corporateclean');
 	$slideshow_pause=theme_get_setting('slideshow_pause','corporateclean');
